==> WaterDemos/02a-Python-in-5-minutes.ipynb <==
# Python in 5 minutes...
While its absurd to think we can learn Python in 5 minutes, it's useful to at least introduce a few basic concepts of the language before we dive in. After covering these, you'll at least get past that first speed bump and can start interacting on your own.

## Variable assignment and data types
Python assigns values to variables with `=`:

zip_code = 27705
myName = "John"
voltage = 9.25

Really the only rules in assigning variables is that the names can't have spaces or begin with numbers. Oh, and there are a few reserved words you can't use, e.g. `class`...

You can check the value of a variable with the `print` command:

print(myName)

When a variable is assigned (or re-assigned) a value, it assumes the **data type** of that value. The most common data types are <u>integers</u>, <u>strings</u>, and <u>float</u>. The data type of a variable can be revealed with the `type` command:

type(myName)

type(zip_code)

<font color=purple>► What's the type of the `voltage` variable?</font>

type()

Data types are importamt because it defines what you can do with that variable. This is because Python is an **object oriented language** meaning everything in Python is an object, and every object has a set of **properties** and **methods** (or defined actions). 

---
## ♦ Python syntax
### Comments
One important concept in Python syntax is how to write comments in your code, i.e., code that included in your script, but that is not run. In Python any code written after a `#` is ignored by the kernel. In the example below, the first line is a comment; it is ingored, but is useful to humans reading the script. 

#Assign my name to the variable myName
myName = 'John'

### Whitespace (indenting) and code blocks
As a scripting language, it's useful to string together multiple statements and run them together. A simple script is just a sequence of commands run chronologically, but you'll likely want to compose more complex scripts that include **conditional execution** (e.g. `if`...`then`... statements) or **loops**, and these require the language to be able to define specific **blocks** of code. In Python, code blocks are identified via **whitespace** or indenting lines of code. Additionally, a colon (`:`) is often used in the line preceeding a code block.<br>Here are a few examples:

for i in range(5):
    x = 2 ** i
    print(x)
print("Finished!")

x = 100
if x < 10:
    print("It's under 10")
elif x < 100:
    print("It's under 100")
else:
    print("It's greater than or equal to 100")

---
## ♦ Python collections
In analyzing data, we deal with multiple related values. Here are typical Python data types for dealing with collections of values.

### Lists
A Python **list** is a collection of values, and the values can vary in their data type. Lists are termed **vectors** because the order of the items it important: it's the only way we can identify specific elements in a list. Python lists are created using brackets: `[`...`]`.

#Create a list
myList = ['Ralph',30,'canoeing',45.3]
#Add one more item to our list
myList.append('Telephone')

To extract elements from a list, use it's **index**. The one catch here is that **Python indices always start at zero, not 1**

#Print the 3rd item in the list
print (myList[2])

### Tuples
A **tuple** is exactly like a list with the one exception is that, once created, it is **immutable**, meaning we can't add or remove items to the list. Python tuples are created using parentheses: `(`...`)`.

#Create the tuple
myTuple = ('Ralph',30,'canoeing',45.3)

#Extract the 1st item from the tuple
myTuple[]

### Dictionaries
A **dictionary** is a collection of items (of any data type) just like a list or a tuple. However, in a dictionary, values are not referenced by their index, but by a **key** we assign to the value when the dictionary is created or when a value is added to the dictionary. Dictionaries are created using curly braces: `{`<u>key</u>:<u>value</u>`}`.

#Create a dictionary, using names as keys
myDict = {'John':21,'Lauren':20,'Martin':30}

print (myDict['Martin'])

myDict['Martin'] = 45
print (myDict['Martin'])

---
## ♦ Python Libraries (for data analytics...)
Python can be extended by importing **libraries** (aka **packages** or **modules**). You may see various forms of importing a library:
* `import numpy`: imports the entire *NumPy* module; NumPy specific statements are preceded with `numpy.`, e.g. `numpy.ndarray(2)`
* `import numpy as np`: as above, but references to Numpy functions are made with `np.`, e.g. `np.ndarray(2)`
* `from numpy import *`: imports the entire *Numpy* module, but functions don't require a prefix: `ndarray(2)`

import numpy as np

np.ndarray(2)

Importing libraries add new **data types**, i.e. new object definitions, and thus new **properties** and **methods** (functions) to our script. The two big data analytics libraries in Python are **NumPy** *(**Num**eric **Py**thon)* and **Pandas** *(**Pan**el **da**ta)*. 

### <u>NumPy</u> and the *n-dimensional array*
Briefly, *NumPy* brings multidimensional analysis to Python with it's key data type: the **n-dimensional array**. This is another collection object, **but** all values in the collection must be the same data type (e.g. all integers or all floats), and indices can assume multiple dimensions. These *ndarrays* offer very speedy calculations and manipulations. 

### <u>Pandas</u> and the *dataframe*
The other big libraru in data analytics, Pandas, actually builds off of NumPy's objects, and introduces the **Dataframe**, which is so central to data analysis and this workshop that we'll examine this on its own. 

## More info
A great book on Numpy and Pandas is available online!<b<>
https://jakevdp.github.io/PythonDataScienceHandbook/index.html
